app/python_exec/requirements.txt
cassandra-driver
pymongo 
virtualenv
